cloud/functions/discovery/requirements.txt
functions-framework>=3.5.0
google-cloud-storage>=2.14.0
requests>=2.31.0
